+++ pyproject.toml
@@ -15,8 +15,14 @@ dependencies = [
     "shapely (>=2.1.1,<3.0.0)",
     "tqdm (>=4.67.1,<5.0.0)",
     "ipykernel (>=6.29.5,<7.0.0)",
-    "matplotlib (>=3.10.3,<4.0.0)",
-    "osmnx (>=2.0.4,<3.0.0)"
+    "matplotlib (>=3.10.3,<4.0.0)"
+]
+
+[project.optional-dependencies]
+dev = [
+    "osmnx (>=2.0.4,<3.0.0)",
+    "h3 (>=3.7.6,<4.0.0)",
+    "scikit-learn (>=1.4.1,<2.0.0)"
 ]
 
 [tool.poetry]
